+++ components/Common/Navigation/index.js
@@ -4,12 +4,12 @@ import styled from "styled-components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faListCheck } from "@fortawesome/free-solid-svg-icons";
 import { faBoxArchive } from "@fortawesome/free-solid-svg-icons";
+import { faHouse } from "@fortawesome/free-solid-svg-icons";
 
 export default function Navigation() {
-  const [selectedNavItem, setSelectedNavItem] = useLocalStorageState(
-    "projects",
-    { defaultValue: "projects" }
-  );
+  const [selectedNavItem, setSelectedNavItem] = useLocalStorageState("home", {
+    defaultValue: "home",
+  });
 
   function handleNavOnClick(event) {
     event.preventDefault();
@@ -20,13 +20,20 @@ export default function Navigation() {
     <Nav>
       <NavItem
         onClick={(event) => handleNavOnClick(event)}
+        className="home"
+        selectedNavItem={selectedNavItem}
+      >
+        <FontAwesomeIcon icon={faHouse}> Home</FontAwesomeIcon>
+        <StyledLink href="/">Home</StyledLink>
+      </NavItem>
+      <NavItem
+        onClick={(event) => handleNavOnClick(event)}
         className="projects"
         selectedNavItem={selectedNavItem}
       >
         <FontAwesomeIcon icon={faListCheck}>Projects</FontAwesomeIcon>
-        <StyledLink href="/">Projects</StyledLink>
+        <StyledLink href="/project">Projects</StyledLink>
       </NavItem>
-
       <NavItem
         onClick={(event) => handleNavOnClick(event)}
         className="stocks"
