+++ components/Common/SubHeading/index.js
@@ -1,6 +1,4 @@
 import styled from "styled-components";
 
-const SubHeading = styled.h2`
-  text-align: center;
-`;
+const SubHeading = styled.p``;
 export default SubHeading;
